Fix and new e2e for (#4387)
